cuny-1969-chatbot-demo/requirements.txt
streamlit
requests
beautifulsoup4
sentence-transformers
chromadb
pandas
pillow
numpy<2.0